+++ src/levelTransitState.cpp
@@ -60,7 +60,7 @@ StateEnum LevelTransitState::run() {
 
     while (running) {
         int currentTime = SDL_GetTicks();
-        int dt = currentTime - lastTime;
+        int dt = (int) (currentTime - lastTime);
         lastTime = currentTime;
         timeElapsed+=dt;
